requirements.txt
streamlit==1.31.0
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2
pydub==0.25.1